refactor: 将所有ID类型从Integer迁移为Long
统一项目中实体、DTO、控制器和数据库层的ID类型，将Integer替换为Long以支持更大的数据范围，修复类型不匹配问题，同时添加批量任务处理的模拟失败测试方法。

+++ spring-whale-core/src/main/java/io/github/springwhale/framework/core/context/AuthenticationContext.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class AuthenticationContext {
-    private Integer userId;
+    private Long userId;
     private String username;
-    private Integer tenantId;
+    private Long tenantId;
 }
