+++ spring-whale-database/src/main/java/io/github/springwhale/database/SimpleBaseEntity.java
@@ -23,7 +23,7 @@ public abstract class SimpleBaseEntity {
     private LocalDateTime createTime;
 
     @UserIdField
-    private Integer createBy;
+    private Long createBy;
 
     @PrePersist
     public void prePersist() {
